working on vite style integration
