+++ public/manifest.json
@@ -21,15 +21,6 @@
     "notifications"
   ],
   "background": {
-    "service_worker": "js/background.js"
-  },
-  "commands": {
-    "_execute_browser_action": {
-      "suggested_key": {
-        "default": "Ctrl+Shift+F",
-        "mac": "MacCtrl+Shift+F"
-      },
-      "description": "Opens popup.html"
-    }
+    "service_worker": "background.js"
   }
 }
