+++ public/options.html
@@ -6,7 +6,6 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="./favicon.ico">
     <title>{{BET}} Options</title>
-    <link rel="stylesheet" href="css/app.css">
   </head>
   <body>
     <noscript>
@@ -20,7 +19,7 @@
       </div>
     </div>
     <footer>
-      <script type="module" src="./js/settings.js"></script>
+      <script type="module" src="./settings.js"></script>
     </footer>
   </body>
 </html>
