+++ public/popup.html
@@ -6,7 +6,6 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="./favicon.ico">
     <title>{{BET}}</title>
-    <link rel="stylesheet" href="css/app.css">
   </head>
   <body>
     <noscript>
@@ -20,7 +19,7 @@
       </div>
     </div>
     <footer>
-      <script type="module" src="./js/app.js"></script>
+      <script type="module" src="./app.js"></script>
     </footer>
   </body>
 </html>
