+++ tooling.tsconfig.json
@@ -1,6 +1,9 @@
 {
     "compilerOptions": {
         "target": "ESNext",
+        "lib": [
+            "ESNext"
+        ],
         "module": "ESNext",
         "outDir": "./tools/",
         "rootDir": "./tools/",
