--- tools/clean.ts
@@ -1,9 +1,0 @@
-import * as fs from "fs";
-
-function removeExport(file: string) {
-	let content = fs.readFileSync(file, "utf8");
-	content = content.replace("export {};", "");
-	fs.writeFileSync(file, content);
-}
-
-removeExport("./dist/js/background.js");
