--- vite.config.settings.js
@@ -1,29 +1,0 @@
-import { defineConfig } from "vite";
-import { resolve } from "path";
-import tsconfigPaths from "vite-tsconfig-paths";
-
-export default defineConfig({
-	build: {
-		rollupOptions: {
-			input: resolve(__dirname, "src/settings.ts"),
-			output: {
-				entryFileNames: "settings.js",
-				dir: resolve(__dirname, "dist/js"),
-			},
-		},
-		sourcemap: true,
-	},
-	plugins: [tsconfigPaths()],
-	resolve: {
-		extensions: [".tsx", ".ts"],
-	},
-	esbuild: {
-		include: /.*\.tsx?$/,
-		exclude: [
-			/node_modules/,
-			/dist/,
-			/src\/background\.ts/,
-			/src\/app\.ts/,
-		],
-	},
-});
